systemd bug fix for stop state
the `disable` and `stop` state are different states for systemd. So
adjust or systemd service so we have two seperate bundles:
 * disable
 * stop

Fixed a bug in the `mpf` framework when we want to stop a service
it does not detect the `activating` state. So the service keeps on
retarting by systemd.